refactor: group main form components in /forms
